+++ src/main/java/id/kedukasi/core/repository/wilayah/KecamatanRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KecamatanRepository extends JpaRepository<MasterKecamatan,Integer> {
+public interface KecamatanRepository extends JpaRepository<MasterKecamatan,Long> {
 }
