+++ src/main/java/id/kedukasi/core/repository/wilayah/KotaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KotaRepository extends JpaRepository<MasterKota,Integer> {
+public interface KotaRepository extends JpaRepository<MasterKota,Long> {
 
 }
