+++ src/main/java/id/kedukasi/core/service/CalonPesertaService.java
@@ -9,20 +9,20 @@ import java.util.Date;
 public interface CalonPesertaService {
     Result getAllCalonPeserta(String uri);
 
-    Result getCalonPesertaById(long id, String uri);
+    Result getCalonPesertaById(Long id, String uri);
 
     ResponseEntity<?> updateCalonPeserta(Long id, Long kelasId, String namaPeserta, Date tanggalLahir,
                                          String jenisKelamin, String pendidikanTerakhir, String noHp, String email,
-                                         MultipartFile uploadImage, Integer provinsi, Integer kota, Integer kecamatan,
-                                         Integer kelurahan, String alamatRumah, String motivasi, String kodeReferal);
+                                         MultipartFile uploadImage, Long provinsi, Long kota, Long kecamatan,
+                                         Long kelurahan, String alamatRumah, String motivasi, String kodeReferal);
 
-    ResponseEntity<?> deleteCalonPeserta(boolean banned, long id, String uri);
+    ResponseEntity<?> deleteCalonPeserta(boolean banned, Long id, String uri);
 
-    ResponseEntity<?> changeToPeserta(long id, String uri);
+    ResponseEntity<?> changeToPeserta(Long id, String uri);
 
-    ResponseEntity<?> changeStatusTes(Long statusTesOrd, long id, String uri);
+    ResponseEntity<?> changeStatusTes(Long statusTesOrd, Long id, String uri);
 
-    ResponseEntity<?> changeKelas(long calonPesertaId, long kelasId, String uri);
+    ResponseEntity<?> changeKelas(Long calonPesertaId, Long kelasId, String uri);
 
     Result filterByStatusTes(Long statusTesOrd);
